Create psx_shader_godot_shadows_and_extras
Not the best fix for the shadows issue but it works. This shader will show lighting stronger though, so I added an extra param  called light_intensity. If you need to dim down the intensity of the light recieved just use a value below 1.0, for instance: 0.5 will half the intensity.

Also added some extra params for roughness and specular, just a quality of life thing if you need to make a metalic material or something more reflective.

The issue with lighting was pretty much that this shader is not rendering correctly the vertex lighting, so the easy fix was to just add the function light() and manually code the lighting. You can play around with DIFFUSE_LIGHT and get cool effects as well. 

Cheers.